lab_8/src/consumer_requirements.txt
cassandra-driver
fastapi
pandas
uvicorn
kafka-python

